Make defaultPageSize a const and move emptyStateComponent into src/components/EmptyState
